Add a rv64_generic_pte ASM function that we jal to, the benefit of that is that it is a generic function for all instances of LEAF's and BRANCH'es in the Sv39 pagetables.  With adjustments this can be furthered to Sv48 or higher.
Right now it duplicates functionality to the original locore.S and I have
put a lot of attention to alignments.  It also differs in some way in that
it will create an L1 entry for the last 1GB under VM_MAX_KERNEL_ADDRESS
of which we have 32 GB (of kernel space).  Originally this was mapped to be
2 * L2_SIZE (2 MiB) but in order to keep gigabyte alignments on gigabytes and
megabyte alignments on megabytes I have chosen it to be this way.

Not tested on other machines than the still broken Mango Pi.