Attempts to get the serial console going.  I think it needs interrupts for the FIFO to work.  so in parallel work on the plic.  We'll port over a FreeBSD driver from Julien Cassette that is a wrapper for PLIC, because the Allwinner D1 Mango Pi has some limitations with regards to priorities and enables on interrupts.  In particular they introduced a permission register for the C906 that disallows supervisor access to PLIC_CTRL, PLIC_PRIO, PLIC_IP, and PLIC_IE registers.  So I was fooling around this yesterday a bit, and by accident had the freebsd diff in my browser open, I felt right at home with what he was writing in the commit log.  So I will port this over.
The com_fdt.c changes are temporary, I was messing with every single possible
combination on this driver it would not work and probably because of no PLIC.

Further more, I was afraid we had the wrong frequency on the APB1 clock, so I
overhauled the sxiccmu.c code for the D1.  I like this better the way it is
now so I'll keep it.

Thankfully I have a Starfive Visionfive 2 now and compiling is tremendously
sped up.  I saw Jeremie's commit for speeding up pmap.c and we're on a week
before his changes, I don't know if I'll merge first or after.